requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Configuration and serialization
pyyaml==6.0.1
python-dotenv==1.0.0

# HTTP client for LLM and MCP communication
aiohttp==3.9.1
httpx==0.25.2

# Async support
asyncio-mqtt==0.16.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Security
cryptography==42.0.8

# Utilities
python-multipart==0.0.6
jinja2==3.1.2 